Move deprecation to common module
